chat_bot/local_test/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0

# IA Générative - APIs Cloud
openai==1.12.0
anthropic==0.18.1

# IA Générative - Local (optionnel)
transformers==4.36.2
torch==2.1.2
accelerate==0.25.0

# Traitement de texte et NLP
nltk==3.8.1
spacy==3.7.2

# Utilities
python-dateutil==2.8.2
colorama==0.4.6
